Added support for some mobile browsers
Newer than 2014, older than 2021 (Chrome For Android) or 2022 (Firefox For Android). 

In the meanwhile, Controls support for Audio was partial and yet to determine. A warning message and an alternate control mean is provided for resolution below 800px width. Especially for old Android devices which can not always update their newer-than-2014 but older-than-2021web browser.