Regression BUGFIX currently untested (read full commit comment)
A short while ago, a typo introduced a regression bug in the auto-recover of pages (if htmlcache enabled), causing the 6-seconds delay let to the user to read about the ongoing recovery of misgenerated page (typically, stored in cache while the media tier was suffering an overload and therefore, missing one or more track titles, having an empty playlist (very unlikely to happen since playlists are cached on front-end and media tiers not often asked for a fresh one)...
...To be reduced to a zero-second delay ; then, not letting the user the ability to act in one of the two following ways:

-either not waiting for six seconds, and asking for a page re-generation immediately
-or canceling the auto-repair and stay on the misgenerated page ; maybed to close the overload splash and take a look at what has gone wrong.

This regression bug was introduced very recently. It should now be cured. It is probable that nobody deployed any server instance incorporating this bug.